src/app/core/guards/not-logged-in.guard.ts
import { AuthService } from './../auth/auth.service';
import { Injectable } from '@angular/core';
import {
  CanActivate,
  ActivatedRouteSnapshot,
  RouterStateSnapshot,
  UrlTree,
  CanActivateChild,
  Router,
} from '@angular/router';
import { Observable } from 'rxjs';

@Injectable({
  providedIn: 'root',
})
export class NotLoggedInGuard implements CanActivateChild {
  constructor(private auth: AuthService, private router: Router) {}

  canActivateChild(
    next: ActivatedRouteSnapshot,
    state: RouterStateSnapshot
  ):
    | Observable<boolean | UrlTree>
    | Promise<boolean | UrlTree>
    | boolean
    | UrlTree {
    // Router protected by this router can only be accessed by not logged in users.
    // If the user is already authenticated redirect them to the home page (/tweets/)

    if (this.auth.isAuthenticated()) {
      this.router.navigate(['tweets']);
      return false;
    }
    return true;
  }
}
